Redefined ConstructFromHex to take float values

+++ VectorFour.cpp
@@ -52,10 +52,10 @@ namespace Vector4
 	}
 
 	void Vector4::ConstructFromHex(
-		int a_iRED,
-		int a_iGREEN,
-		int a_iBLUE,
-		int a_iALPHA
+		float a_iRED,
+		float a_iGREEN,
+		float a_iBLUE,
+		float a_iALPHA
 		)
 	{	
 		fW = a_iRED   / 255;
